Add Expenses & Test Cases

+++ src/main/java/invoice/Invoice.java
@@ -5,9 +5,9 @@ import java.util.Date;
 public class Invoice {
     private Long id;
     private Date date;
-    private double totalValue;
+    private float totalValue;
 
-    public Invoice(Long id, Date date, double totalValue) {
+    public Invoice(Long id, Date date, float totalValue) {
         this.id = id;
         this.date = date;
         this.totalValue = totalValue;
@@ -29,11 +29,11 @@ public class Invoice {
         this.date = date;
     }
 
-    public double getTotalValue() {
+    public float getTotalValue() {
         return totalValue;
     }
 
-    public void setTotalValue(double totalValue) {
+    public void setTotalValue(float totalValue) {
         this.totalValue = totalValue;
     }
 }
